fix(ddb): use two instances, select primary as read instance
We previously went to one instance thinking that there was a problem
with stale reads. This turned out to be CloudFront caching api traffic.

With a single instance and no secondaries, mongoose tries forever to
reach the secondary, in vain.

This should be solved by using `readPreference=primary` in the conn
string, but to allow for longevity of the setup, we will bump ddb back
to 2 instances.

We will also specify a read preference of `primaryPreferred`, to
better assure some consistency